MOVE: describe to pandas/io/formats/describe.py
